requirements.txt
tkinter
sv-ttk
python-vlc
ffmpeg-python